feat(premium): /premium をランチャー化し TOPIX33 を /premium/market へ分離
- /premium をグラフ無しのランチャー(機能カード導線)へ全面刷新
- 旧 /premium のヒートマップ+業種比較チャート+要約を /premium/market へ移設(専用ページなので直接表示)
- portfolio を loadPortfolio() seam 経由にし、将来のサーバ保存へ画面非破壊で移行できるようにする
- 不要になった CollapsibleSection を削除
- 再設計の判断を decision-log に記録
